pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adscan"
version = "5.0.0"
description = "ADscan launcher (host-side) that orchestrates Docker to run the ADscan runtime."
readme = "adscan_pypi_wrapper/README.md"
authors = [{ name = "ADscan" }]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "rich>=13.0,<14.0",
    "requests>=2.31.0,<3.0",
    "packaging>=23.0,<26.0",
    "certifi>=2023.0.0",
    "sentry-sdk>=1.40.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.9.0,<1.0",
    "build>=1.2.0,<2.0",
    "twine>=6.0.0,<7.0",
    "mypy>=1.8.0",
    "types-requests",
    "pytest-html",
    "jinja2",
]
cli = [
    "psutil",
    "python-docx",
    "python-magic>=0.4.27,<0.5; platform_system == 'Linux'",
    "markitdown>=0.1.5,<0.2; python_version >= '3.10'",
    "pydantic-ai>=0.8.1",
    "netifaces",
    "prompt_toolkit",
    "questionary",
    "neo4j>=5.0.0",
    "posthog",
    "pypsrp",
    "structlog>=24.1.0",
    "pydantic-settings>=2.4.0",
]

[project.urls]
homepage = "https://github.com/ADscanPro/adscan"

[project.scripts]
adscan = "adscan_launcher.cli:main"
"adscan.py" = "adscan_launcher.cli:main"

[tool.setuptools]
packages = ["adscan_core", "adscan_launcher"]
